pyproject.toml
[project]
name = "kaggle-cmi"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "numpy>=1.25.2",
    "pandas>=2.0.3",
    "polars>=0.19",
    "jupyter>=1.0.0",
    "matplotlib>=3.7.2",
    "seaborn>=0.12.2",
    "transformers>=4.31.0",
    "pillow==9.5.0",
    "torch==2.0.1+cu118",
    "torchvision==0.15.2+cu118",
    "torchaudio==2.0.2+cu118",
    "lightgbm>=4.0.0",
    "catboost>=1.2",
    "xgboost>=1.7.6",
    "hydra-core>=1.3.2",
    "hydra_colorlog>=1.2.0",
    "google-cloud-storage>=2.10.0",
    "deepspeed>=0.10.0",
    "seqeval>=1.2.2",
    "openmim>=0.3.9",
    "fairscale>=0.4.13",
    "wandb>=0.15.8",
    "pytorch-lightning==2.0.6",
    "timm==0.9.2",
    "pyarrow>=14.0.1",
    "matplotlib-venn>=0.11.9",
    "scipy>=1.11.3",
    "numba>=0.58.1",
    "loguru>=0.7.2",
    "segmentation_models_pytorch>=0.3.3",
    "invoke>=2.2.0",
    "pathlib>=1.0.1",
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["*"]

[tool.rye]
managed = true
dev-dependencies = [
    "black>=23.7.0",
    "mypy>=1.5.0",
    "flake8>=6.1.0",
    "ruff>=0.0.284",
    "kaggle>=1.5.16",
]

[tool.hatch.metadata]
allow-direct-references = true

[[tool.rye.sources]]
name = "torch"
url = "https://download.pytorch.org/whl/cu118"
type = "index"

[[tool.rye.sources]]
name = "torchvision"
url = "https://download.pytorch.org/whl/cu118"
type = "index"

[[tool.rye.sources]]
name = "cuml-cu11"
url = "https://pypi.nvidia.com"
type = "index"


[tool.ruff]
target-version = "py311"
line-length = 100
ignore = [
    "E402",   # Module level import not at top of file
    "E501",   # Line too long
]
select = [
    "F",
    "E",
    "W",
    "I",
    "B",
]
